fix: participantCount change int to float
참가자 수를 Int로 했을 때 발생하는 오류로 인해서 float로 변경하였다

+++ Assets/PrivateFolder/HyunJune/Script/Vote/VoteManager.cs
@@ -68,7 +68,7 @@ public class VoteManager : MonoBehaviourPun
 
     public bool deadBodyFinder = false;
     private bool voteComplete = false;
-    private int participantCount;
+    private float participantCount;
     private VoteRole myRole;
 
     [Header("SkipWindow")]
@@ -425,7 +425,7 @@ public class VoteManager : MonoBehaviourPun
         VotingResult();
     }
 
-    public int CalculateAlivePlayer()
+    public float CalculateAlivePlayer()
     {
         int alivePlayerCount = 0;
         foreach (KeyValuePair<int, Photon.Realtime.Player> player in PhotonNetwork.CurrentRoom.Players)
